split run function out to separate file
