+++ runProcessRainfall.py
@@ -1,0 +1,6 @@
+from processRainfall import runAPICall
+import time
+
+while(True):
+    runAPICall(None,None)
+    time.sleep(600)
